nufft grid dtypes

+++ src/aspire/basis/fle_2d.py
@@ -261,7 +261,7 @@ class FLEBasis2D(SteerableBasis2D, FBBasisMixin):
         self.num_angular_nodes = num_angular_nodes
 
         # create gridpoints
-        nodes = 1 - (2 * np.arange(self.num_radial_nodes) + 1) / (
+        nodes = 1 - (2 * np.arange(self.num_radial_nodes, dtype=self.dtype) + 1) / (
             2 * self.num_radial_nodes
         )
         nodes = (np.cos(np.pi * nodes) + 1) / 2
@@ -276,8 +276,8 @@ class FLEBasis2D(SteerableBasis2D, FBBasisMixin):
         phi = (
             2 * np.pi * np.arange(self.num_angular_nodes // 2) / self.num_angular_nodes
         )
-        x = np.cos(phi).reshape(1, self.num_angular_nodes // 2)
-        y = np.sin(phi).reshape(1, self.num_angular_nodes // 2)
+        x = np.cos(phi, dtype=self.dtype).reshape(1, self.num_angular_nodes // 2)
+        y = np.sin(phi, dtype=self.dtype).reshape(1, self.num_angular_nodes // 2)
         x = x * nodes * h
         y = y * nodes * h
         self.grid_x = x.flatten()
